Factor duplicate keyboard stuff into shared
Also clean up some files that should have been deleted.
