+++ druid-shell/src/platform/mod.rs
@@ -21,12 +21,15 @@ cfg_if::cfg_if! {
     } else if #[cfg(target_os = "macos")] {
         mod mac;
         pub use mac::*;
+        pub(crate) mod shared;
     } else if #[cfg(all(feature = "x11", target_os = "linux"))] {
         mod x11;
         pub use x11::*;
+        pub(crate) mod shared;
     } else if #[cfg(target_os = "linux")] {
         mod gtk;
         pub use self::gtk::*;
+        pub(crate) mod shared;
     } else if #[cfg(target_arch = "wasm32")] {
         mod web;
         pub use web::*;
